:hammer: 엔티티 매핑 개선 AdminService Test

+++ src/main/java/com/giyong/community/controller/MainCategoryController.java
@@ -44,7 +44,7 @@ public class MainCategoryController {
     }
 
     @GetMapping("/categorys")
-    public String categories(Integer mainCategoryId, Integer page, Model m) {
+    public String categories(Long mainCategoryId, Integer page, Model m) {
         MainCategory mainCategory = mainCategoryService.findById(mainCategoryId);
         m.addAttribute("mainCategory", mainCategory);
         m.addAttribute("page", page);
