+++ src/main/java/com/giyong/community/repository/MainCategoryRepository.java
@@ -3,6 +3,6 @@ package com.giyong.community.repository;
 import com.giyong.community.entity.MainCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MainCategoryRepository extends JpaRepository<MainCategory, Integer> {
+public interface MainCategoryRepository extends JpaRepository<MainCategory, Long> {
     MainCategory findByMainCategoryName(String mainCategoryName);
 }
